+++ display_metrics_ios/ios/display_metrics_ios/Package.swift
@@ -1,0 +1,28 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "display_metrics_ios",
+    platforms: [
+        .iOS("12.0"),
+    ],
+    products: [
+        .library(name: "display-metrics-ios", targets: ["display_metrics_ios"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "display_metrics_ios",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ],
+            resources: [
+                .process("PrivacyInfo.xcprivacy"),
+            ]
+        )
+    ]
+)